requirements.txt
numpy~=1.23.5
pandas~=1.5.2
hydra-core~=1.3.1
hydra-optuna-sweeper
omegaconf~=2.3.0
scipy~=1.9.3
scikit-learn~=1.1.3
tqdm~=4.64.1
pyg~=2.2.0
pytorch~=1.13.1
pytorch-cuda~=11.7
mlflow[extras]
